rename ign* files to gz
