+++ rmoss_gz_base/src/gz_module/gz_chassis_actuator.cpp
@@ -11,12 +11,12 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#include "rmoss_ign_base/ign_chassis_actuator.hpp"
+#include "rmoss_gz_base/gz_chassis_actuator.hpp"
 
 #include <memory>
 #include <string>
 
-namespace rmoss_ign_base
+namespace rmoss_gz_base
 {
 
 
@@ -42,4 +42,4 @@ void IgnChassisActuator::set(const geometry_msgs::msg::Twist & data)
   ign_chassis_cmd_pub_->Publish(ign_msg);
 }
 
-}  // namespace rmoss_ign_base
+}  // namespace rmoss_gz_base
