+++ rmoss_gz_cam/src/gz_cam_main.cpp
@@ -15,12 +15,12 @@
 #include <memory>
 
 #include "rclcpp/rclcpp.hpp"
-#include "rmoss_ign_cam/ign_cam_node.hpp"
+#include "rmoss_gz_cam/gz_cam_node.hpp"
 
 int main(int argc, char * argv[])
 {
   rclcpp::init(argc, argv);
-  auto node = std::make_shared<rmoss_ign_cam::IgnCamNode>();
+  auto node = std::make_shared<rmoss_gz_cam::GzCamNode>();
   rclcpp::spin(node->get_node_base_interface());
   rclcpp::shutdown();
   return 0;
